+++ javascript/modules/client/src/internal/client-core.ts
@@ -314,6 +314,100 @@ export abstract class ClientCore implements IClientCore {
       });
   }
 
+  protected estimateDeposit(params: IDeposit): Promise<IGasFeeEstimation> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const [
+      daoAddress,
+      amount,
+      tokenAddress,
+      reference,
+    ] = ClientCore.createDepositParameters(params);
+
+    const daoInstance = DAO__factory.connect(daoAddress, this.connectedSigner);
+
+    const override =
+      tokenAddress !== AddressZero
+        ? {}
+        : {
+            value: amount,
+          };
+
+    const gasLimit = daoInstance.estimateGas.deposit(
+      tokenAddress,
+      amount,
+      reference,
+      override
+    );
+    return this.estimateGasFee(gasLimit);
+  }
+
+  protected increaseAllowance(
+    tokenAddress: string,
+    daoAddress: string,
+    amount: bigint
+  ): Promise<void> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const governanceERC20Instance = GovernanceERC20__factory.connect(
+      tokenAddress,
+      this.connectedSigner
+    );
+
+    return governanceERC20Instance
+      .increaseAllowance(daoAddress, BigNumber.from(amount))
+      .then(tx => tx.wait())
+      .then(cr => {
+        if (
+          BigNumber.from(amount).gt(
+            cr.events?.find(e => e?.event === "Approval")?.args?.value
+          )
+        ) {
+          throw new Error("Could not increase allowance");
+        }
+      });
+  }
+
+  protected estimateIncreaseAllowance(
+    tokenAddress: string,
+    daoAddress: string,
+    amount: bigint
+  ): Promise<IGasFeeEstimation> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const governanceERC20Instance = GovernanceERC20__factory.connect(
+      tokenAddress,
+      this.connectedSigner
+    );
+
+    const gasLimit = governanceERC20Instance.estimateGas.increaseAllowance(
+      daoAddress,
+      BigNumber.from(amount)
+    );
+    return this.estimateGasFee(gasLimit);
+  }
+
+  protected currentAllowance(
+    tokenAddress: string,
+    daoAddress: string
+  ): Promise<bigint> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const governanceERC20Instance = GovernanceERC20__factory.connect(
+      tokenAddress,
+      this.connectedSigner
+    );
+
+    return this.connectedSigner
+      .getAddress()
+      .then(address => governanceERC20Instance.allowance(address, daoAddress))
+      .then(allowance => allowance.toBigInt());
+  }
+
   // IPFS METHODS
 
   public async pin(input: string | Uint8Array): Promise<string> {